rest api

+++ workstocks/src/main/java/it/workstocks/service/impl/NewsServiceImpl.java
@@ -5,12 +5,11 @@ import java.util.Optional;
 import java.util.Set;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.cache.annotation.CacheEvict;
-import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -23,13 +22,13 @@ import it.workstocks.dto.pagination.PaginatedResponse;
 import it.workstocks.entity.blog.Comment;
 import it.workstocks.entity.blog.Like;
 import it.workstocks.entity.blog.News;
-import it.workstocks.entity.user.applicant.Applicant;
 import it.workstocks.exception.WorkstocksBusinessException;
-import it.workstocks.repository.blog.CommentRepository;
-import it.workstocks.repository.blog.NewsRepository;
-import it.workstocks.repository.user.ApplicantRepository;
+import it.workstocks.repository.CommentRepository;
+import it.workstocks.repository.NewsRepository;
 import it.workstocks.service.NewsService;
 import it.workstocks.utils.AuthUtility;
+import it.workstocks.utils.ErrorUtils;
+import it.workstocks.utils.Translator;
 
 @Service
 public class NewsServiceImpl implements NewsService{
@@ -38,20 +37,20 @@ public class NewsServiceImpl implements NewsService{
 	private NewsRepository newsRepository;
 	
 	@Autowired
-	private ApplicantRepository applicantRepository;
+	private EntityMapper mapper;
 	
 	@Autowired
-	EntityMapper mapper;
+	private CommentRepository commentRepository;
 	
 	@Autowired
-	CommentRepository commentRepository;
+	private Translator translator;
 	
 	private News findOptionalNews(String id) throws WorkstocksBusinessException {
 		Optional<News> news = newsRepository.findById(id);
 		if (news.isPresent()) {
 			return news.get();
 		} else {
-			throw new WorkstocksBusinessException("News non trovata");
+			throw new WorkstocksBusinessException(translator.toLocale(ErrorUtils.NOT_FOUND, new Object[] {"news", id}),HttpStatus.NOT_FOUND);
 		}
 	}
 	
@@ -66,6 +65,7 @@ public class NewsServiceImpl implements NewsService{
 		paginatedReponse.setTotalElements(newsPaginated.getTotalElements());
 		paginatedReponse.setPageNumber(newsPaginated.getPageable().getPageNumber() + 1);
 		paginatedReponse.setTotalPages(newsPaginated.getTotalPages());
+		paginatedReponse.setPageSize(newsPaginated.getNumberOfElements());
 		PaginatedDtoResponse<NewsDto> response = new PaginatedDtoResponse<>();
 		response.setResponse(paginatedReponse); 
 		response.setElements(content);
@@ -74,8 +74,8 @@ public class NewsServiceImpl implements NewsService{
 	}
 	
 	@Override
-	public PaginatedDtoResponse<NewsDto> findAllPaginatedTitleFiltering(int pageNumber, String titleFilter) throws WorkstocksBusinessException {
-		Pageable pageable = PageRequest.of(pageNumber - 1, 9, getSortByCreationDate());
+	public PaginatedDtoResponse<NewsDto> findAllPaginatedTitleFiltering(Integer pageNumber, Integer pageSize, String titleFilter) {
+		Pageable pageable = PageRequest.of(pageNumber - 1, pageSize, getSortByCreationDate());
 		
 		Page<News> newsPaginated = null;
 		if (titleFilter == null || titleFilter.trim().isEmpty()) {
@@ -88,44 +88,45 @@ public class NewsServiceImpl implements NewsService{
 	}
 	
 	@Override
-	public PaginatedDtoResponse<NewsDto> findAllNewsPaginatedAdmin(int pageNumber) throws WorkstocksBusinessException {
-		Pageable pageable = PageRequest.of(pageNumber - 1, 10, getSortByCreationDate());
-		Page<News> newsPaginated = newsRepository.findAll(pageable);
-		return getPaginatedNewsDto(newsPaginated);
+	public int countNewsComments(String newsId) {
+		return commentRepository.countByNewsId(newsId);
 	}
 	
-	private void extendComments(Set<CommentDto> comments) throws WorkstocksBusinessException {
+	/*private void extendComments(Set<CommentDto> comments) throws WorkstocksBusinessException {
 		if (comments != null) {
 			for (CommentDto comment : comments) {
 				Optional<Applicant> user = applicantRepository.findById(comment.getUserId());
-				if (user.isEmpty()) throw new WorkstocksBusinessException("Applicant comment not found");
+				if (user.isEmpty()) throw new WorkstocksBusinessException("Applicant for comment not found", HttpStatus.NOT_FOUND);
 				comment.setUserName(user.get().getName());
-				comment.setUserAvatar(user.get().getAvatar());
+				comment.convertBase64UserAvatar(user.get().getAvatar());
 			}
 		}
+	}*/
+	
+	@Override
+	public byte[] findPhotoById(String id) throws WorkstocksBusinessException {
+		News news = findOptionalNews(id);
+		return news.getImage();
 	}
 
 	@Override
-	public NewsDto findById(String id, boolean isAdminDetail) throws WorkstocksBusinessException {
+	public NewsDto findById(String id) throws WorkstocksBusinessException {
 		NewsDto news = mapper.toDto(findOptionalNews(id));
-		// se admin che richiede la news per la modifica non prendo i commenti
-		if (!isAdminDetail) {
-			news.setCommentSize(commentRepository.countByNewsId(news.getId()));
-			extendComments(news.getRecentComments());
-		}
-		
+		news.setCommentSize(commentRepository.countByNewsId(news.getId()));
 		return news;
 	}
 	
 	@Override
-	public PaginatedDtoResponse<CommentDto> findCommentsPaginatedByNewsId(String newsId, int pageNumber) throws WorkstocksBusinessException {
+	public PaginatedDtoResponse<CommentDto> findCommentsPaginatedByNewsId(String newsId, Integer pageNumber, Integer limit) throws WorkstocksBusinessException {
+		findOptionalNews(newsId);
+		
 		PaginatedResponse paginatedReponse = new PaginatedResponse();
 		paginatedReponse.setPageNumber(pageNumber);
 		
-		Pageable pageable = PageRequest.of(pageNumber - 1, News.MAX_RECENT_COMMENT_SIZE, getSortByCreationDate());
+		Pageable pageable = PageRequest.of(pageNumber - 1, limit, getSortByCreationDate());
 		Page<Comment> commentsPaginated = commentRepository.findByNewsId(newsId, pageable);
 		Set<CommentDto> content = mapper.toDto(new LinkedHashSet<>(commentsPaginated.getContent()));
-		extendComments(content);
+		//extendComments(content);
 		paginatedReponse.setTotalElements(commentsPaginated.getTotalElements());
 		paginatedReponse.setTotalPages(commentsPaginated.getTotalPages());
 		
@@ -136,46 +137,40 @@ public class NewsServiceImpl implements NewsService{
 		return response;
 	}
 	
-	// admin
-
-	@Override
-	@CacheEvict(value = "footerNews", allEntries = true)
-	@Transactional(rollbackFor = WorkstocksBusinessException.class)
-	public void upsertNews(NewsDto newsDto) throws WorkstocksBusinessException {
-		News news = null;
-		if (newsDto.getId() != null) {
-			news = findOptionalNews(newsDto.getId());
-			mapper.updateNewsEntity(newsDto, news);
-		} else {
-			news = mapper.toEntity(newsDto);
-		}		
-		
-		newsRepository.save(news);
-	}
-
-	@Override
-	@CacheEvict(value = "footerNews", allEntries = true)
-	@Transactional(rollbackFor = WorkstocksBusinessException.class)
-	public void deleteNewsById(String id) throws WorkstocksBusinessException {
-		newsRepository.deleteById(id);
-		commentRepository.deleteAllByNewsId(id);
-	}
-	
 	// applicant
+	
+	@Override
+	public boolean checkUserLike(String newsId) throws WorkstocksBusinessException {
+		News news = findOptionalNews(newsId);
+		Long userId = AuthUtility.getCurrentApplicant().getId();
+		return news.getLikes().stream().anyMatch(like -> like.getUserId().equals(userId));
+	}
 
 	@Override
-	@CacheEvict(value = "footerNews", allEntries = true)
 	@Transactional(rollbackFor = WorkstocksBusinessException.class)
-	public void addOrRemoveLike(String newsId, boolean isAdding) throws WorkstocksBusinessException {
+	public void addLike(String newsId) throws WorkstocksBusinessException {
 		News news = findOptionalNews(newsId);
 		Long userId = AuthUtility.getCurrentApplicant().getId();
 		
-		if (isAdding) {
-			Like like = new Like();
-			like.setUserId(userId);
-			news.getLikes().add(like);
-		} else {
-			news.getLikes().removeIf(like -> like.getUserId().equals(userId));
+		if (news.getLikes().stream().anyMatch(like -> like.getUserId().equals(userId))) {
+			throw new WorkstocksBusinessException(translator.toLocale(ErrorUtils.LIKE_ALREADY_SUBMIT, new Object[] {newsId, userId}),HttpStatus.CONFLICT);
+		}
+		
+		Like like = new Like();
+		like.setUserId(userId);
+		news.getLikes().add(like);		
+		newsRepository.save(news);
+	}
+	
+	@Override
+	@Transactional(rollbackFor = WorkstocksBusinessException.class)
+	public void removeLike(String newsId) throws WorkstocksBusinessException {
+		News news = findOptionalNews(newsId);
+		Long userId = AuthUtility.getCurrentApplicant().getId();
+		boolean isPresent = news.getLikes().removeIf(like -> like.getUserId().equals(userId));
+		
+		if (!isPresent) {
+			throw new WorkstocksBusinessException(translator.toLocale(ErrorUtils.LIKE_NOT_FOUND, new Object[] {newsId, userId}),HttpStatus.NOT_FOUND);
 		}
 		
 		newsRepository.save(news);
@@ -183,11 +178,13 @@ public class NewsServiceImpl implements NewsService{
 
 	@Override
 	@Transactional(rollbackFor = WorkstocksBusinessException.class)
-	public void deleteComment(String commentId) throws WorkstocksBusinessException {
+	public void deleteComment(String commentId, String newsId) throws WorkstocksBusinessException {
+		findOptionalNews(newsId);
+		
 		Optional<Comment> comment = commentRepository.findById(commentId);
-		if (comment.isPresent()) {
+		if (comment.isPresent() && comment.get().getNews().getId().equals(newsId)) {
 			if (!AuthUtility.compareCurrentUser(comment.get().getUserId())) {
-				throw new AccessDeniedException("User not authorized to delete comment");
+				throw new AccessDeniedException(String.format("User with id %d not authorized to delete comment with id %s", comment.get().getUserId(), commentId));
 			}
 			commentRepository.delete(comment.get());
 			
@@ -202,30 +199,25 @@ public class NewsServiceImpl implements NewsService{
 			}
 			
 		} else {
-			throw new WorkstocksBusinessException("Comment not found");
+			throw new WorkstocksBusinessException(translator.toLocale(ErrorUtils.COMMENT_NOT_FOUND, new Object[] {commentId, newsId}),HttpStatus.NOT_FOUND);
 		}
 	}
 
 	@Override
 	@Transactional(rollbackFor = WorkstocksBusinessException.class)
-	public void addComment(CommentDto commentDto, String newsId) throws WorkstocksBusinessException {
+	public String addComment(CommentDto commentDto, String newsId) throws WorkstocksBusinessException {
 		News news = findOptionalNews(newsId);
 		
-		Comment comment = mapper.toEntity(commentDto);
+		Comment comment = new Comment();
+		comment.setBody(commentDto.getBody());
 		comment.setUserId(AuthUtility.getCurrentApplicant().getId());
 		comment.setNews(news);
-		commentRepository.save(comment);
+		comment = commentRepository.save(comment);
 		
 		news.addComment(comment);
 		newsRepository.save(news);
-	}
-
-	@Override
-	@Cacheable(value = "footerNews")
-	public Set<NewsDto> findLatests(int numberOfNews) throws WorkstocksBusinessException {
-		Pageable pageable = PageRequest.of(0, numberOfNews, getSortByCreationDate());
-		Page<News> latestNews = newsRepository.findAll(pageable);
-		return mapper.toDtoNewsCollection(new LinkedHashSet<>(latestNews.getContent()));
+		
+		return comment.getId();
 	}
 
 }
